+++ src/materials/metal.h
@@ -7,11 +7,11 @@
 
 class Metal : public Material {
 public:
-	Metal(const Color& albedo, double fuzz);
+	Metal(const Color& albedo, float fuzz);
 
 	bool scatter(const Ray& r, const HitRecord& record, Color& attenuation_out, Ray& scattered_out) const override;
 
 private:
 	Color _albedo;
-	double _fuzz;
+	float _fuzz;
 };
