+++ src/math/interval.cpp
@@ -4,12 +4,12 @@
 
 Interval Interval::empty{MathUtility::infinity, -MathUtility::infinity};
 Interval Interval::universe{-MathUtility::infinity, MathUtility::infinity};
-Interval Interval::unit{0.0, 1.0};
+Interval Interval::unit{0.f, 1.f};
 
 Interval::Interval() : min(MathUtility::infinity), max(-MathUtility::infinity) {
 }
 
-Interval::Interval(double p_min, double p_max) : min(p_min), max(p_max) {
+Interval::Interval(float p_min, float p_max) : min(p_min), max(p_max) {
 }
 
 Interval::Interval(const Interval& a, const Interval& b) {
@@ -17,30 +17,30 @@ Interval::Interval(const Interval& a, const Interval& b) {
 	max = a.max >= b.max ? a.max : b.max;
 }
 
-void Interval::expand(double delta) {
+void Interval::expand(float delta) {
 	*this = this->expanded(delta);
 }
 
-Interval Interval::expanded(double delta) const {
-	auto padding = delta / 2.0;
+Interval Interval::expanded(float delta) const {
+	auto padding = delta / 2.f;
 	return {min - padding, max + padding};
 }
 
-double Interval::clamped(double x) const {
+float Interval::clamped(float x) const {
 	if(x < min) return min;
 	if(x > max) return max;
 
 	return x;
 }
 
-double Interval::size() const {
+float Interval::size() const {
 	return max - min;
 }
 
-bool Interval::contains(double x) const {
+bool Interval::contains(float x) const {
 	return min <= x && x <= max;
 }
 
-bool Interval::surrounds(double x) const {
+bool Interval::surrounds(float x) const {
 	return min < x && x < max;
 }
